Refactor CompactSplashPotionMessage to be in chat
